models/engine: DB: More useful methods for db
In the current setup there is one problem. we have to make several db calls in case we need attributes from diffrent types of objects.

This `join` method could be very useful for making `JOIN` queries using relationships between the tables. This is just an excuse to avoid some db calls

Signed-off-by: ZorEl212 <yabsirad212@gmail.com>