models: Handle deleting/creating users properly
When a user is deleted, it's db data is still preserved for statical purposes, because of that when a new user is created with the same `linux_username` later on it causes conflicts for several reasons.

1. `linux_username` is a unique attribute on the db, so even if we can create a system user with that name, we cannot register that entry on the db.

2. Who are we to say someone won't request a linux account that has been used before by another user, in such case the user gotta come up with different name.

So, to counter such cases introduce a new flag to identify an active user (an entry with an active rental and system user) stale user (an entry available just for the sake of stats)

This way there can be ONLY ONE entry that has this flag disabled.